refactor: applied most of the comments regarding the naming and error handling
