+++ Makefile
@@ -5,8 +5,8 @@ lint:
 	golangci-lint run ./...
 
 build:
-	go build -o bin/ginserver ./cmd/ginmain/main.go
-	go build -o bin/httpserver ./cmd/httpmain/main.go
+	go build -o bin/ginserver ./cmd/gin_server/main.go
+	go build -o bin/httpserver ./cmd/http_server/main.go
 
 docker:
 	docker compose up
