+++ gin.Dockerfile
@@ -6,7 +6,7 @@ COPY . .
 
 RUN go mod download
 
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o ginserver cmd/ginmain/main.go 
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o ginserver cmd/gin_server/main.go 
 
 FROM gcr.io/distroless/base
