+++ http.Dockerfile
@@ -6,7 +6,7 @@ COPY . .
 
 RUN go mod download
 
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o httpserver cmd/httpmain/main.go 
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o httpserver cmd/http_server/main.go 
 
 FROM gcr.io/distroless/base
