+++ pkg/ginserver/server.go
@@ -1,6 +1,7 @@
 package ginserver
 
 import (
+	"net/http"
 	"time"
 
 	"github.com/gin-gonic/gin"
@@ -12,11 +13,11 @@ func GinHandler() *gin.Engine {
 	router.GET("/datetime", func(c *gin.Context) {
 		contenttype := c.GetHeader("Content-Type")
 		if contenttype == "application/json" {
-			c.JSON(200, gin.H{
+			c.JSON(http.StatusOK, gin.H{
 				"datetime": time.Now().Local().Format("Monday 02-01-2006 15:04:05"),
 			})
 		} else {
-			c.String(200, time.Now().Local().Format("Monday 02-01-2006 15:04:05"))
+			c.String(http.StatusOK, time.Now().Local().Format("Monday 02-01-2006 15:04:05"))
 
 		}
 	})
